:wrench: refactor: restructure layout components and move sections to separate folders - refs #19
