+++ src/App.tsx
@@ -1,4 +1,4 @@
-import { Home } from '@pages/Home';
+import { Home } from '@pages/home';
 
 export function App() {
   return (
