--- src/components/carousel/index.ts
@@ -1,2 +1,0 @@
-export * from './carousel';
-export * from './carouselSlideNavigation';
