--- src/components/footer/index.ts
@@ -1,5 +1,0 @@
-export * from './footer';
-export * from './footerInfo';
-export * from './footerSection';
-export * from './footerLinkButton';
-export * from './footerButtonGroup';
