+++ src/components/layout/footer/index.ts
@@ -1,0 +1,5 @@
+export * from './footer';
+export * from './info';
+export * from './section';
+export * from './link-button';
+export * from './button-group';
