+++ src/components/layout/footer/info.tsx
