+++ src/components/sections/carousel/index.ts
@@ -1,0 +1,2 @@
+export * from './carousel';
+export * from './slide-navigation';
