+++ src/components/sections/carousel/slide-navigation.tsx
