+++ src/components/sections/hero.tsx
